+++ src/FluentlyHttpClient/FluentHttpClientOptions.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
+using System.Net.Http.Formatting;
+using FluentlyHttpClient.Middleware;
 
 namespace FluentlyHttpClient
 {
@@ -20,7 +22,7 @@ namespace FluentlyHttpClient
 		public TimeSpan Timeout { get; set; }
 
 		/// <summary>
-		/// Gets or sets the identifier (key) for the http client.
+		/// Gets or sets the identifier (key) for the HTTP client.
 		/// </summary>
 		public string Identifier { get; set; }
 
@@ -32,16 +34,21 @@ namespace FluentlyHttpClient
 		/// <summary>
 		/// Gets or sets the middleware to be used for each request.
 		/// </summary>
-		public List<Type> Middleware { get; set; }
+		public List<MiddlewareConfig> Middleware { get; set; }
 
 		/// <summary>
-		/// Handler to customize request on creation. In order to specify defaults as desired, or so.
+		/// Gets or sets handler to customize request on creation. In order to specify defaults as desired, or so.
 		/// </summary>
 		public Action<FluentHttpRequestBuilder> RequestBuilderDefaults { get; set; }
 
 		/// <summary>
-		/// HTTP handler stack to use for sending requests.
+		/// Gets or sets HTTP handler stack to use for sending requests.
 		/// </summary>
 		public HttpMessageHandler HttpMessageHandler { get; set; }
+
+		/// <summary>
+		/// Gets or sets formatters to be used for content negotiation, for "Accept" and body media formats. e.g. JSON, XML, etc...
+		/// </summary>
+		public MediaTypeFormatterCollection Formatters { get; set; }
 	}
 }