+++ src/FluentlyHttpClient/ServiceCollectionExtensions.cs
@@ -6,12 +6,12 @@ using FluentlyHttpClient.Middleware;
 namespace Microsoft.Extensions.DependencyInjection
 {
 	/// <summary>
-	/// Extension methods for setting up fluent http services in an <see cref="T:Microsoft.Extensions.DependencyInjection.IServiceCollection" />.
+	/// Extension methods for setting up fluent HTTP services in an <see cref="T:Microsoft.Extensions.DependencyInjection.IServiceCollection" />.
 	/// </summary>
 	public static class FluentlyHttpClientServiceCollectionExtensions
 	{
 		/// <summary>
-		/// Adds fluently http client services to the specified <see cref="IServiceCollection"/>.
+		/// Adds fluently HTTP client services to the specified <see cref="IServiceCollection"/>.
 		/// </summary>
 		/// <param name="services"></param>
 		/// <returns>Returns service collection for chaining.</returns>